@@ -9,21 +9,31 @@
     "deploy": "npm run build && npx gh-pages -d build"
   },
   "devDependencies": {
-    "@fontsource/fira-mono": "^4.5.10",
-    "@neoconfetti/svelte": "^1.0.0",
-    "@playwright/test": "^1.28.1",
-    "@sveltejs/adapter-auto": "^2.0.0",
-    "@sveltejs/adapter-static": "^2.0.3",
-    "@sveltejs/kit": "^1.20.4",
-    "@types/cookie": "^0.5.1",
-    "autoprefixer": "^10.4.16",
-    "postcss": "^8.4.30",
-    "postcss-load-config": "^4.0.1",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^5.0.0",
-    "vite": "^4.4.2"
+    "@nvl/sveltex": "^0.4.3",
+    "@sveltejs/adapter-static": "^3.0.0",
+    "@sveltejs/kit": "^2.0.0",
+    "@sveltejs/package": "^2.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@types/eslint": "^9.6.0",
+    "autoprefixer": "^10.4.21",
+    "daisyui": "^5.0.27",
+    "eslint": "^9.7.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.36.0",
+    "gh-pages": "^6.3.0",
+    "postcss": "^8.5.3",
+    "prettier": "^3.3.2",
+    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "publint": "^0.2.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
+    "sveltekit-html-minifier": "^1.0.4",
+    "tailwindcss": "^3.4.17",
+    "vite": "^5.0.11"
   },
-  "type": "module"
+  "type": "module",
+  "peerDependencies": {
+    "svelte": "^3.x || ^4.x"
+  }
 }
